all done!

+++ src/app/modules/notations/components/update-one/update-one.component.ts
@@ -1,9 +1,10 @@
 import {Component, OnInit} from '@angular/core';
 import {Router} from "@angular/router";
-import {INotation} from "../../models/notation";
+import {INotation} from "../../models";
 import {FormControl, FormGroup, Validators} from "@angular/forms";
-import {NotationsService} from "../../services/notations.service";
-import {DataService} from "../../services/data.service";
+import {DataService, NotationsService} from "../../services";
+import {DialogComponent} from "../dialog/dialog.component";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   selector: 'app-update-one',
@@ -12,51 +13,88 @@ import {DataService} from "../../services/data.service";
 })
 export class UpdateOneComponent implements OnInit {
 
-  notation: INotation
-  updatedNotation: INotation
+  notation: INotation;
+  updatedNotation: INotation;
 
+  // Forms controls
   controls = {
     title: new FormControl('', [Validators.required]),
     descriptions: new FormControl('', [Validators.required])
   }
 
-
+  // Form
   form: FormGroup = new FormGroup(this.controls);
 
   constructor(
     private router: Router,
     private notationsService: NotationsService,
-    private dataService: DataService
+    private dataService: DataService,
+    private dialog: MatDialog
   ) {
-    // Get selected notation from state
 
+    // Get selected notation from state
     this.notation = this.router.getCurrentNavigation()?.extras.state as INotation;
     // Fill forms inputs
-    this.controls.title.reset(this.notation.title)
-    this.controls.descriptions.reset(this.notation.descriptions)
+    this.controls.title.reset(this.notation.title);
+    this.controls.descriptions.reset(this.notation.descriptions);
   }
 
   ngOnInit(): void {
 
   }
 
+  // Clean old information from inputs functions
+  cleanInputs() {
+    this.controls.title.setValue('');
+    this.controls.descriptions.setValue('');
+  }
+
   saveChanges() {
-    // Save value from form in updatedNotation
-    this.updatedNotation = {
-      title: this.form.controls.title.value,
-      descriptions: this.form.controls.descriptions.value,
-      createdAt: this.notation.createdAt
-    }
+    // Confirm action in dialog
+    const dialogRef = this.dialog.open(DialogComponent);
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
 
-    //Remove outdated notation
-    this.notationsService.removeNotation(this.notation)
+        // Save value from form in updatedNotation
+        this.updatedNotation = {
+          title: this.form.controls.title.value,
+          descriptions: this.form.controls.descriptions.value,
+          createdAt: this.notation.createdAt
+        }
 
-    //Add updated notation to localStore and data.services
-    const updatedNotationsList = this.notationsService.addNotation(this.updatedNotation)
-    this.dataService.setNotationsList(updatedNotationsList)
+        //Remove outdated notation
+        this.notationsService.removeNotation(this.notation);
 
-    // Clean old information
-    this.controls.title.setValue('')
-    this.controls.descriptions.setValue('')
+        //Add updated notation to localStore and data.services
+        const updatedNotationsList = this.notationsService.addNotation(this.updatedNotation);
+        this.dataService.setNotationsList(updatedNotationsList);
+
+        // Clean old information
+        this.cleanInputs();
+      }
+    });
+  }
+
+  deleteNotation() {
+    // Confirm action in dialog
+    const dialogRef = this.dialog.open(DialogComponent);
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        // Delete from localStore and data.service
+        this.dataService.setNotationsList(this.notationsService.removeNotation(this.notation));
+        this.cleanInputs();
+      }
+    });
+  }
+
+  cancelChanges() {
+    // Confirm action in dialog
+    const dialogRef = this.dialog.open(DialogComponent);
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.controls.title.reset(this.notation.title);
+        this.controls.descriptions.reset(this.notation.descriptions);
+      }
+    });
   }
 }
